Allow 'changefreq' to be specified for each page
The sitemap.xml format contains an optional element "changefreq" which,
if given, tells the search-engine robots approximately how often a given
page will change. For example, the front page of a site might change
daily, but archived blog posts probably never change. This branch checks
to see whether a "changefreq" has been specified in a file's YAML front
matter and, if so, this is added to the sitemap.xml. (If no changefreq
is specified then the element is omitted from the XML.)